refactor(app): move theme creation out of main.tsx into src/theme/
- theme/index.ts (design tokens) renamed to theme/tokens.ts; index.ts now
  re-exports tokens + the composed theme so existing imports keep working
- New theme/palette.ts, theme/typography.ts, theme/components.ts,
  theme/create-theme.ts — same values as before (palette hexes now reference
  the gray/background tokens they duplicated)
- New src/app.tsx provider composition (ThemeProvider > CssBaseline >
  AppRouter); main.tsx is bootstrap-only
- test-utils.tsx now renders with the real app theme instead of a bare
  createTheme(), so tests exercise the production overrides
- Smoke test for the composed theme

Part 3 of the frontend modernization roadmap.
