+++ app/src/app.tsx
@@ -1,0 +1,14 @@
+import CssBaseline from '@mui/material/CssBaseline';
+import { ThemeProvider } from '@mui/material/styles';
+
+import { AppRouter } from 'src/router';
+import { theme } from 'src/theme';
+
+export function App() {
+  return (
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      <AppRouter />
+    </ThemeProvider>
+  );
+}
