+++ app/src/theme/typography.ts
@@ -1,0 +1,7 @@
+import type { ThemeOptions } from '@mui/material/styles';
+
+import { typography as fontStacks } from 'src/theme/tokens';
+
+export const typography: ThemeOptions['typography'] = {
+  fontFamily: fontStacks.fontFamily,
+};
